// convex/_generated/api.d.ts
/* eslint-disable */
/**
 * Generated `api` utility.
 *
 * THIS CODE IS AUTOMATICALLY GENERATED.
 *
 * To regenerate, run `npx convex dev`.
 * @module
 */

import type * as assets from "../assets.js";
import type * as auth from "../auth.js";
import type * as authHelpers from "../authHelpers.js";
import type * as contact from "../contact.js";
import type * as executive from "../executive.js";
import type * as http from "../http.js";
import type * as labs from "../labs.js";
import type * as meetings from "../meetings.js";
import type * as migration from "../migration.js";
import type * as news from "../news.js";
import type * as pila from "../pila.js";
import type * as rbac from "../rbac.js";
import type * as storage from "../storage.js";
import type * as users from "../users.js";

import type {
  ApiFromModules,
  FilterApi,
  FunctionReference,
} from "convex/server";

declare const fullApi: ApiFromModules<{
  assets: typeof assets;
  auth: typeof auth;
  authHelpers: typeof authHelpers;
  contact: typeof contact;
  executive: typeof executive;
  http: typeof http;
  labs: typeof labs;
  meetings: typeof meetings;
  migration: typeof migration;
  news: typeof news;
  pila: typeof pila;
  rbac: typeof rbac;
  storage: typeof storage;
  users: typeof users;
}>;

/**
 * A utility for referencing Convex functions in your app's public API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = api.myModule.myFunction;
 * ```
 */
export declare const api: FilterApi<
  typeof fullApi,
  FunctionReference<any, "public">
>;

/**
 * A utility for referencing Convex functions in your app's internal API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = internal.myModule.myFunction;
 * ```
 */
export declare const internal: FilterApi<
  typeof fullApi,
  FunctionReference<any, "internal">
>;

export declare const components: {};